requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pandas
numpy

openai

langchain
langchain-openai
langchain-community
chromadb
tiktoken